Merged PR 2005: Add shutdown process that is testable. Fix some race conditions.
Allow the final kill_process function to be overridden
with an event raise that a test can look out for.
Add a separate config timeout value for this for overriding. Only place this is actually done is in the tests however.

Fix race conditions in http.py.
1) get_ai_statuses_for_api can fail if the training_lookup dictionary is modified during iteration.
I noticed glitches in the build lab during test runs, which I think is this.

2) Kibana is showing multiple initialization of
the server in multiple log entries at same millisecond timestamp. Looks like we're not
following the proper documented on_startup method. Add check to protect against double initialization

Related work items: #3656